Reworked dice
Dice no longer change background color based on selection.
Dice now change icon to an alternate picture with a yellow background.
Implementation notes:
ImageIcon[] pics is now 2D ImageIcon[][], where ImageIcon[0] is an ImageIcon[] for the white pics, and ImageIcon[1] is an ImageIcon[] for the yellow pics.
Changed the way the rollUnselected() method checks for which buttons are selected. --> Now checks through the white row in pics, and compares to the icon of the button.
Fixed a bug where the scores would not display after selecting a button.