Slight Changes and a Menu
Fixed slight bugs, minor changes to GUI.

Added a Menu Bar that Gives the Option to Restart the Game as a 1 or 2 player game. Names of players are reset to "Player 1" and "Player 2" or "Computer" (Based on the # of players). Also added an option on the menu that opens a new window where you can set the player's names. The layout of everything is just a first draft, tell me any suggestions for formatting.