Changed the pop-up name options
Now a JDialog instead of JFrame. This allows me to disable the main game while the options are open, and the main game is not able to be focused on or clicked until the user closes the pop-up. Without it, I could open an infinite amount of option windows. (Not Ideal)